+++ python/acl_anthology/people/__init__.py
@@ -13,7 +13,14 @@
 # limitations under the License.
 
 from .name import Name, NameSpecification, ConvertableIntoName
-from .person import Person
+from .person import Person, NameLink
 from .index import PersonIndex
 
-__all__ = ["ConvertableIntoName", "Name", "NameSpecification", "Person", "PersonIndex"]
+__all__ = [
+    "ConvertableIntoName",
+    "Name",
+    "NameLink",
+    "NameSpecification",
+    "Person",
+    "PersonIndex",
+]