+++ python/acl_anthology/people/index.py
@@ -14,7 +14,7 @@
 
 from __future__ import annotations
 
-from attrs import define, field, asdict
+from attrs import define, field
 from collections.abc import Iterable
 from collections import Counter, defaultdict
 import itertools as it
@@ -22,7 +22,7 @@ from pathlib import Path
 from rich.progress import track
 from scipy.cluster.hierarchy import DisjointSet  # type: ignore
 import sys
-from typing import cast, Any, TYPE_CHECKING
+from typing import cast, Any, Optional, TYPE_CHECKING
 import yaml
 
 try:
@@ -31,10 +31,15 @@ except ImportError:  # pragma: no cover
     from yaml import Loader, Dumper  # type: ignore
 
 from ..containers import SlottedDict
-from ..exceptions import AnthologyException, AmbiguousNameError, NameIDUndefinedError
-from ..utils.ids import AnthologyIDTuple
+from ..exceptions import (
+    AnthologyException,
+    NameSpecResolutionError,
+    PersonDefinitionError,
+)
+from ..utils.ids import AnthologyIDTuple, is_verified_person_id
 from ..utils.logging import get_logger
-from . import Person, Name, NameSpecification
+from . import Person, Name, NameLink, NameSpecification
+from .name import _YAMLName
 
 if TYPE_CHECKING:
     from _typeshed import StrPath
@@ -42,7 +47,7 @@ if TYPE_CHECKING:
     from ..collections import Paper, Volume
 
 log = get_logger()
-VARIANTS_FILE = "yaml/name_variants.yaml"
+PEOPLE_INDEX_FILE = "yaml/people.yaml"
 
 
 @define
@@ -62,18 +67,31 @@ class PersonIndex(SlottedDict[Person]):
     Attributes:
         parent: The parent Anthology instance to which this index belongs.
         verbose: If False, will not show progress bar when building the index from scratch.
-        name_to_ids: A mapping of [Name][acl_anthology.people.name.Name] instances to person IDs.
+        path: The path to `people.yaml`.
+        by_orcid: A mapping of ORCIDs (as strings) to person IDs.
+        by_name: A mapping of [Name][acl_anthology.people.name.Name] instances to lists of person IDs.
+        slugs_to_verified_ids: A mapping of strings (representing slugified names) to lists of person IDs.
         similar: A [disjoint-set structure][scipy.cluster.hierarchy.DisjointSet] of persons with similar names.
         is_data_loaded: A flag indicating whether the index has been constructed.
     """
 
     parent: Anthology = field(repr=False, eq=False)
     verbose: bool = field(default=True)
-    name_to_ids: dict[Name, list[str]] = field(
+    path: Path = field(init=False)
+    # TODO: could the following fields be made private and have getters that check for self.is_data_loaded?
+    by_orcid: dict[str, str] = field(init=False, repr=False, default={})
+    by_name: dict[Name, list[str]] = field(
+        init=False, repr=False, factory=lambda: defaultdict(list)
+    )
+    slugs_to_verified_ids: dict[str, list[str]] = field(
         init=False, repr=False, factory=lambda: defaultdict(list)
     )
     similar: DisjointSet = field(init=False, repr=False, factory=DisjointSet)
     is_data_loaded: bool = field(init=False, repr=True, default=False)
+
+    @path.default
+    def _path(self) -> Path:
+        return self.parent.datadir / Path(PEOPLE_INDEX_FILE)
 
     def get_by_name(self, name: Name) -> list[Person]:
         """Access persons by their name.
@@ -86,12 +104,12 @@ class PersonIndex(SlottedDict[Person]):
         """
         if not self.is_data_loaded:
             self.load()
-        return [self.data[pid] for pid in self.name_to_ids[name]]
+        return [self.data[pid] for pid in self.by_name[name]]
 
     def get_by_namespec(self, name_spec: NameSpecification) -> Person:
         """Access persons by their name specification.
 
-        See [get_or_create_person()][acl_anthology.people.index.PersonIndex.get_or_create_person] for exceptions that can be raised by this function.
+        See [resolve_namespec()][acl_anthology.people.index.PersonIndex.resolve_namespec] for exceptions that can be raised by this function.
 
         Parameters:
             name_spec: A name specification.
@@ -101,7 +119,22 @@ class PersonIndex(SlottedDict[Person]):
         """
         if not self.is_data_loaded:
             self.load()
-        return self.get_or_create_person(name_spec, create=False)
+        return self.resolve_namespec(name_spec)
+
+    def get_by_orcid(self, orcid: str) -> Person | None:
+        """Access persons by their ORCID.
+
+        Parameters:
+            orcid: A string representing an ORCID.
+
+        Returns:
+            The person with that ORCID, if it exists, otherwise None.
+        """
+        if not self.is_data_loaded:
+            self.load()
+        if orcid in self.by_orcid:
+            return self.data[self.by_orcid[orcid]]
+        return None
 
     def find_coauthors(
         self, person: str | Person, include_volumes: bool = True
@@ -143,13 +176,9 @@ class PersonIndex(SlottedDict[Person]):
                 and not cast("Volume", item).has_frontmatter
             ):
                 continue
-            coauthors.update(
-                self.get_or_create_person(ns, create=False).id for ns in item.editors
-            )
+            coauthors.update(self.resolve_namespec(ns).id for ns in item.editors)
             if hasattr(item, "authors"):
-                coauthors.update(
-                    self.get_or_create_person(ns, create=False).id for ns in item.authors
-                )
+                coauthors.update(self.resolve_namespec(ns).id for ns in item.authors)
         del coauthors[person.id]
         return coauthors
 
@@ -164,7 +193,9 @@ class PersonIndex(SlottedDict[Person]):
     def reset(self) -> None:
         """Resets the index."""
         self.data = {}
-        self.name_to_ids = defaultdict(list)
+        self.by_orcid = {}
+        self.by_name = defaultdict(list)
+        self.slugs_to_verified_ids = defaultdict(list)
         self.similar = DisjointSet()
         self.is_data_loaded = False
 
@@ -175,8 +206,7 @@ class PersonIndex(SlottedDict[Person]):
             Exceptions raised during the index creation are sent to the logger, and only a generic exception is raised at the end.
         """
         self.reset()
-        # Load variant list, so IDs defined there are added first
-        self._load_variant_list()
+        self._load_people_index()
         # Go through every single volume/paper and add authors/editors
         iterator = track(
             self.parent.collections.values(),
@@ -190,7 +220,7 @@ class PersonIndex(SlottedDict[Person]):
                 context: Paper | Volume = volume
                 try:
                     for name_spec in volume.editors:
-                        person = self.get_or_create_person(name_spec)
+                        person = self.resolve_namespec(name_spec, allow_creation=True)
                         person.item_ids.append(volume.full_id_tuple)
                     for paper in volume.papers():
                         context = paper
@@ -202,9 +232,9 @@ class PersonIndex(SlottedDict[Person]):
                             else paper.get_editors()
                         )
                         for name_spec in name_specs:
-                            person = self.get_or_create_person(name_spec)
+                            person = self.resolve_namespec(name_spec, allow_creation=True)
                             person.item_ids.append(paper.full_id_tuple)
-                except Exception as exc:
+                except Exception as exc:  # pragma: no cover
                     note = f"Raised in {context.__class__.__name__} {context.full_id}; {name_spec}"
                     # If this is merged into a single if-statement (with "or"),
                     # the type checker complains ¯\_(ツ)_/¯
@@ -220,6 +250,51 @@ class PersonIndex(SlottedDict[Person]):
             )
         self.is_data_loaded = True
 
+    def _load_people_index(self) -> None:
+        """Load and parse the `people.yaml` file.
+
+        Raises:
+            KeyError: If `people.yaml` contains a malformed person ID; or if a person is listed without any names.
+        """
+        merge_list: list[tuple[str, str]] = []
+
+        with open(self.path, "r", encoding="utf-8") as f:
+            data = yaml.load(f, Loader=Loader)
+
+        for pid, entry in data.items():
+            if not is_verified_person_id(pid):
+                raise KeyError(
+                    f"Invalid person ID in people.yaml: {pid}"
+                )  # pragma: no cover
+            self.add_person(
+                Person(
+                    id=pid,
+                    parent=self.parent,
+                    names=[Name.from_dict(n) for n in entry.pop("names")],
+                    orcid=entry.pop("orcid", None),
+                    comment=entry.pop("comment", None),
+                    degree=entry.pop("degree", None),
+                    similar_ids=entry.get("similar", []),
+                    disable_name_matching=entry.pop("disable_name_matching", False),
+                    is_explicit=True,
+                )
+            )
+            for similar_id in entry.pop("similar", []):
+                merge_list.append((pid, similar_id))
+
+            # Check for unprocessed keys to catch errors
+            if entry:
+                log.warning(
+                    f"people.yaml: entry '{pid}' has unknown keys: {entry.keys()}"
+                )  # pragma: no cover
+
+        # Process IDs with similar names
+        for pid_list in self.slugs_to_verified_ids.values():
+            for pid in pid_list[1:]:
+                self.similar.merge(pid_list[0], pid)
+        for a, b in merge_list:
+            self.similar.merge(a, b)
+
     def add_person(self, person: Person) -> None:
         """Add a new person to the index.
 
@@ -230,93 +305,146 @@ class PersonIndex(SlottedDict[Person]):
             raise KeyError(f"A Person with ID '{pid}' already exists in the index")
         self.data[pid] = person
         self.similar.add(pid)
+        if person.orcid is not None:
+            self.by_orcid[person.orcid] = pid
         for name in person.names:
-            self.name_to_ids[name].append(pid)
+            self.by_name[name].append(pid)
+            if is_verified_person_id(pid):
+                self.slugs_to_verified_ids[name.slugify()].append(pid)
 
-    def get_or_create_person(
-        self, name_spec: NameSpecification, create: bool = True
-    ) -> Person:
-        """Get the person represented by a name specification, or create a new one if needed.
+    def ingest_namespec(self, name_spec: NameSpecification) -> NameSpecification:
+        """Update a name specification for ingestion, potentially filling in the ID field.
+
+        If the name specification contains an ORCID but doesn't have an ID yet, this will find the person with this ORCID and fill in their ID; if it doesn't exist yet, it will create a new person with a "verified" ID and fill in the new, generated ID.  The supplied name specification will be modified in-place, but also returned.
 
         Parameters:
             name_spec: The name specification on the paper, volume, etc.
-            create: If False, will not create a new Person object, but instead raise `NameIDUndefinedError` if no person matching `name_spec` exists.  Defaults to True.
 
         Returns:
-            The person represented by `name_spec`.  This will try to use the `id` attribute if it is set, look up the name in the index otherwise, or try to find a matching person by way of an ID clash.  If all of these fail, it will create a new person and return that.
+            The name specification as it should be used for the new ingestion material.
+        """
+        if name_spec.orcid is None or name_spec.id is not None:
+            return name_spec
+
+        if (person := self.get_by_orcid(name_spec.orcid)) is not None:
+            name_spec.id = person.id
+            # Make sure the name used here is listed for this person
+            person.add_name(name_spec.name)
+        else:
+            # Need to create a new person; generate name slug for the ID
+            pid = name_spec.name.slugify()
+            if pid in self.data:
+                # ID is already in use; add last four digits of ORCID to disambiguate
+                pid = f"{pid}-{name_spec.orcid[-4:]}"
+
+            self.add_person(
+                Person(
+                    id=pid,
+                    parent=self.parent,
+                    names=[name_spec.name] + name_spec.variants,
+                    orcid=name_spec.orcid,
+                    is_explicit=True,
+                )
+            )
+            name_spec.id = pid
+
+        return name_spec
+
+    def resolve_namespec(
+        self, name_spec: NameSpecification, allow_creation: bool = False
+    ) -> Person:
+        """Resolve a name specification to a person, potentially creating a new unverified person instance.
+
+        Parameters:
+            name_spec: The name specification on the paper, volume, etc.
+            allow_creation: If True, will instantiate a new Person object with an unverified ID if no person matching `name_spec` exists.  Defaults to False.
+
+        Returns:
+            The person represented by `name_spec`.  If `name_spec.id` is set, this will determine the person to resolve to.  Otherwise, the slugified name will be used to find a matching person; an explicitly-defined (verified) person can be returned if exactly one such person exists and does not have `disable_name_matching` set.  In all other cases, it will resolve to an unverified person.
 
         Raises:
-            AmbiguousNameError: If there are multiple known IDs for the given name, but there is no explicit `id` attribute.
-            NameIDUndefinedError: If there is an explicit `id` attribute, but the ID has not been defined.
+            NameSpecResolutionError: If `name_spec` cannot be resolved to a Person and `allow_creation` is False.
+            PersonDefinitionError: If `name_spec.id` is set, but either the ID or the name used with the ID has not been defined in `people.yaml`. (Inherits from NameSpecResolutionError)
         """
         name = name_spec.name
         if (pid := name_spec.id) is not None:
-            # Explicit ID given; should already exist from name_variants.yaml
-            person = self.data.get(pid)
-            if person is None or not person.is_explicit:
-                exc1 = NameIDUndefinedError(
-                    name_spec, f"Name '{name}' used with ID '{pid}' that doesn't exist"
+            # Explicit ID given – should be explicitly defined in people.yaml
+            if pid not in self.data or not (person := self.data[pid]).is_explicit:
+                raise PersonDefinitionError(
+                    name_spec, f"ID '{pid}' wasn't defined in people.yaml"
                 )
-                exc1.add_note("Did you forget to define the ID in name_variants.yaml?")
-                raise exc1
-            person.add_name(name)
-        elif pid_list := self.name_to_ids[name]:
-            # Name already exists in the index, but has no explicit ID
-            if len(pid_list) > 1:
-                exc2 = AmbiguousNameError(
-                    name,
-                    f"Name '{name.as_first_last()}' is ambiguous, but was used without an ID",
+            if not person.has_name(name):
+                raise PersonDefinitionError(
+                    name_spec,
+                    f"ID '{pid}' was used with name '{name}' that wasn't defined in people.yaml",
                 )
-                exc2.add_note(f"Known IDs are: {', '.join(pid_list)}")
-                raise exc2
-            pid = pid_list[0]
-            person = self.data[pid]
+            if name_spec.orcid is not None and name_spec.orcid != person.orcid:
+                raise PersonDefinitionError(
+                    name_spec,
+                    f"ID '{pid}' was used with ORCID '{name_spec.orcid}', but people.yaml has '{person.orcid}'",
+                )
         else:
-            # Name not in the index and has no explicit ID
-            pid = self.generate_id(name)
-            try:
-                # If the auto-generated ID already exists, we assume it's the same person
-                person = self.data[pid]
-                # If the name scores higher than the current canonical one, we
-                # also assume we should set this as the canonical one
-                if (not person.is_explicit) and (
-                    name.score() > person.canonical_name.score()
-                ):
-                    person.set_canonical_name(name)
-                else:
-                    person.add_name(name)
-                self.name_to_ids[name].append(pid)
-            except KeyError:
-                if create:
-                    # If the auto-generated ID doesn't exist yet, then and only
-                    # then do we create a new person
-                    person = Person(id=pid, parent=self.parent, names=[name])
+            # No explicit ID given
+            if name_spec.orcid is not None:
+                exc1 = NameSpecResolutionError(
+                    name_spec,
+                    "NameSpecification defines an ORCID without an ID",
+                )
+                exc1.add_note(
+                    "To specify an ORCID on a paper, the person needs to have an entry in `people.yaml` and be used with an explicit ID."
+                )
+                raise exc1
+
+            # Generate slug for name matching
+            slug = name.slugify()
+
+            # Check if the slugified name matches any verified IDs
+            matching_ids = self.slugs_to_verified_ids.get(slug, [])
+            if (
+                len(matching_ids) == 1
+                and not (person := self.data[matching_ids[0]]).disable_name_matching
+            ):
+                # Slug unambiguously maps to person and name matching not disabled
+                pid = person.id
+                if not person.has_name(name):
+                    person.add_name(name, inferred=True)
+                    self.by_name[name].append(pid)
+
+            else:
+                # Resolve to unverified ID
+                pid = f"unverified/{slug}"
+
+                if pid in self.data:
+                    # Unverified ID already exists; assume it's the same person
+                    person = self.data[pid]
+                    if not person.has_name(name):
+                        # If the name scores higher than the current canonical
+                        # one, we also assume we should set this as the
+                        # canonical one
+                        if name.score() > person.canonical_name.score():
+                            person.set_canonical_name(name, inferred=True)
+                        else:
+                            person.add_name(name, inferred=True)
+                        self.by_name[name].append(pid)
+                elif allow_creation:
+                    # Unverified ID doesn't exist yet; create it
+                    person = Person(
+                        id=pid, parent=self.parent, names=[(name, NameLink.INFERRED)]
+                    )
                     self.add_person(person)
                 else:
-                    raise NameIDUndefinedError(
+                    raise NameSpecResolutionError(
                         name_spec,
-                        f"Name '{name}' generated ID '{pid}' that doesn't exist",
+                        f"NameSpecification resolved to ID '{pid}' which doesn't exist",
                     )
+
         # Make sure that name variants specified here are registered
         for name in name_spec.variants:
-            person.add_name(name)
-            if name not in self.name_to_ids:
-                self.name_to_ids[name].append(pid)
+            if not person.has_name(name):
+                person.add_name(name, inferred=True)
+            if name not in self.by_name:
+                self.by_name[name].append(pid)
         return person
-
-    @staticmethod
-    def generate_id(name: Name) -> str:
-        """Generates and returns an ID from the given name.
-
-        Warning:
-            This **intentionally doesn't guarantee uniqueness** of the generated ID.
-            If two names generate identical IDs with this method, we assume they
-            refer to the same person.  This happens e.g. when there are missing
-            accents in one version, or when we have an inconsistent first/last split
-            for multiword names.  These cases have in practice always referred to
-            the same person.
-        """
-        return name.slugify()
 
     def _add_to_index(
         self, namespecs: Iterable[NameSpecification], item_id: AnthologyIDTuple
@@ -329,89 +457,36 @@ class PersonIndex(SlottedDict[Person]):
             return
 
         for namespec in namespecs:
-            person = self.get_or_create_person(namespec)
+            person = self.resolve_namespec(namespec, allow_creation=True)
             person.item_ids.append(item_id)
 
-    def _load_variant_list(self) -> None:
-        """Loads and parses the `name_variant.yaml` file.
-
-        Raises:
-            AmbiguousNameError: If there are ambiguous "canonical" names without explicit, unique IDs for each one.
-        """
-        filename = self.parent.datadir / Path(VARIANTS_FILE)
-        merge_list: list[tuple[str, str]] = []
-        with open(filename, "r", encoding="utf-8") as f:
-            variant_list = yaml.load(f, Loader=Loader)
-        for entry in variant_list:
-            # Every entry must have a "canonical" name
-            canonical = Name.from_dict(entry["canonical"])
-            # If it doesn't define an ID, we have to create one
-            if (pid := entry.get("id")) is None:
-                pid = self.generate_id(canonical)
-                if pid in self.data:
-                    raise AmbiguousNameError(
-                        canonical,
-                        (
-                            f"While parsing {filename}: "
-                            f"name '{canonical.as_first_last()}' is ambiguous, but the "
-                            f"automatically generated ID '{pid}' already exists."
-                        ),
-                    )
-            # Parse all the variant names, and make sure canonical stays at index 0
-            names = [canonical] + [
-                Name.from_dict(var) for var in entry.get("variants", [])
-            ]
-            # Now we can create a new person from this entry...
-            person = Person(
-                id=pid,
-                parent=self.parent,
-                names=names,
-                comment=entry.get("comment", None),
-                is_explicit=True,
-            )
-            # ...and add it to the index
-            self.add_person(person)
-            for similar_id in entry.get("similar", []):
-                merge_list.append((pid, similar_id))
-
-        # Process IDs with similar names
-        for name, pid_list in self.name_to_ids.items():
-            for pid in pid_list[1:]:
-                self.similar.merge(pid_list[0], pid)
-        for a, b in merge_list:
-            self.similar.merge(a, b)
-
-    def save(self, path: StrPath) -> None:
-        """Save the entire index.
-
-        CURRENTLY UNTESTED; DO NOT USE.
+    def save(self, path: Optional[StrPath] = None) -> None:
+        """Save the `people.yaml` file.
 
         Arguments:
-            path: The filename to save to.
+            path: The filename to save to. If None, defaults to the parent Anthology's `people.yaml` file.
         """
-        data = []
+        if path is None:
+            path = self.path
+
+        data = {}
         for person in self.values():
+            if not person.is_explicit:
+                continue
+
             attrib: dict[str, Any] = {
-                "id": person.id,
-                "canonical": asdict(
-                    person.canonical_name,
-                    filter=lambda a, v: not (a.name == "script" and v is None),
-                ),
+                "names": [
+                    _YAMLName(name)
+                    for (name, link_type) in person._names
+                    if link_type == NameLink.EXPLICIT
+                ],
+                "comment": person.comment,
+                "degree": person.degree,
+                "disable_name_matching": person.disable_name_matching,
+                "orcid": person.orcid,
+                "similar": person.similar_ids,
             }
-            if person.item_ids:
-                attrib["items"] = person.item_ids
-            if len(person.names) > 1:
-                attrib["variants"] = [
-                    asdict(
-                        name, filter=lambda a, v: not (a.name == "script" and v is None)
-                    )
-                    for name in person.names[1:]
-                ]
-            similar = self.similar.subset(person.id)
-            if len(similar) > 1:
-                attrib["similar"] = [id_ for id_ in similar if id_ != person.id]
-            if person.comment is not None:
-                attrib["comment"] = person.comment
-            data.append(attrib)
+            data[person.id] = {k: v for k, v in attrib.items() if v}
+
         with open(path, "w", encoding="utf-8") as f:
-            yaml.dump(data, f, Dumper=Dumper)
+            yaml.dump(data, f, allow_unicode=True, Dumper=Dumper)